build: native changes — v1.0.16 (30)

+++ android/app/src/main/java/com/testmusicapp/widget/WidgetBridgeModule.kt
@@ -182,7 +182,7 @@ class WidgetBridgeModule(private val reactContext: ReactApplicationContext)
     }
 
     @ReactMethod fun addListener(eventName: String) {}
-    @ReactMethod fun removeListeners(count: Int) {}
+    @ReactMethod fun removeListeners(count: Double) {}
 
     private fun ReadableMap.safeString(key: String): String? =
         if (hasKey(key) && !isNull(key)) getString(key) else null
